+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 4.2.0 (2016-06-XX)
 
+ * Handle the `useAutoItem` option in the URL generator (see #489).
  * Correctly implement the "autoplay" option in the YouTube/Vimeo element (see #509).
  * Show the system messages in a modal dialog (see #486).
  * Make DropZone the default uploader in the file manager (see #504).